chore(structure): move date utils inside its own folder
